Tournament controller actions to support users
Added support for linking tournaments and users. Updated Tournament
controller to support user in required actions.

+++ src/controllers/tournaments.js
@@ -1,6 +1,7 @@
 const tournamentsRouter = require('express').Router();
 
 const Tournament = require('../models/tournament');
+const User = require('../models/user');
 const authenticateJWT = require('../utils/middleware').authenticateJWT;
 
 tournamentsRouter.use(authenticateJWT);
@@ -21,40 +22,54 @@ tournamentsRouter.get('/:id', async (request, response) => {
 });
 
 tournamentsRouter.post('/', async (request, response) => {
+  const user = await User.findById(request.user.id);
+
   const tournament = new Tournament({
-    ...request.body
+    ...request.body,
+    user: user._id
   });
 
   const savedTournament = await tournament.save();
+
+  user.tournaments = user.tournaments.concat(savedTournament._id);
+  await user.save();
 
   return response.status(201).json(savedTournament);
 });
 
 tournamentsRouter.delete('/:id', async (request, response) => {
+  const user = await User.findById(request.user.id);
   const tournament = await Tournament.findById(request.params.id);
 
-  if (tournament === null) {
-    return response.status(404).end();
+  if (tournament == null || tournament.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'Unauthorized request' });
   }
 
   await tournament.deleteOne();
+
+  user.tournaments = user.tournaments.filter(tournament => tournament != null && tournament.toString() !== request.params.id);
+  await user.save();
 
   return response.status(204).end();
 });
 
 tournamentsRouter.put('/:id', async (request, response) => {
+  const user = await User.findById(request.user.id);
+  const tournament = await Tournament.findById(request.params.id);
+
+  if (tournament == null || tournament.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'Unauthorized request' });
+  }
+
   let updatedTournament = {
     name: request.body.name,
     players: request.body.players,
     playerPools: request.body.playerPools,
-    teams: request.body.teams
+    teams: request.body.teams,
+    user: user.id
   };
 
   updatedTournament = await Tournament.findByIdAndUpdate(request.params.id, updatedTournament, { new: true });
-
-  if (updatedTournament === null) {
-    return response.status(404).end();
-  }
 
   return response.json(updatedTournament);
 });
